chore(SFS): rename sfs resources and data sources test files to sfsturbo
