+++ huaweicloud/services/acceptance/sfsturbo/data_source_huaweicloud_sfs_turbo_obs_targets_test.go
@@ -1,4 +1,4 @@
-package sfs
+package sfsturbo
 
 import (
 	"fmt"
