+++ huaweicloud/services/acceptance/sfsturbo/data_source_huaweicloud_sfs_turbos_test.go
@@ -1,4 +1,4 @@
-package sfs
+package sfsturbo
 
 import (
 	"fmt"
