+++ huaweicloud/services/acceptance/sfsturbo/resource_huaweicloud_sfs_turbo_test.go
@@ -1,4 +1,4 @@
-package sfs
+package sfsturbo
 
 import (
 	"fmt"
